pyproject.toml
[project]
name = "whitespace-microservice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "joblib>=1.5.2",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
